# src/aureomall-swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://localhost:3080
...
paths:
...
  /auth/refresh:
    post:
      tags:
          - refresh
      summary: Làm mới accessToken
      description: Làm mới accessToken bằng cookie
      operationId: refresh
      responses:
        '200':
          description: Successfully refreshed access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessAuthenticationRefresh'
        '400':
          description: Refresh failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInvalidErrorResponse'
      security:
        - BearerAuth: []
        - CookieAuth: []
  /users:
    get:
      tags:
          - user
      summary: Hiển thị tất cả người dùng
      description: Hiển thị tất cả người dùng
      operationId: getAllUsers
      responses:
        '200':
          description: getAllUsers successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '400':
          description: 'Token is not valid.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInvalidErrorResponse'
      security:
        - BearerAuth: []
...
components:
  schemas:
...
    SuccessAuthenticationRefresh:
      type: object
      properties:
        successToken:
          type: string
          description: New access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJhZG1pbnByb0BnbWFpbC5jb20iLCJyb2xlX2lkIjoxLCJpYXQiOjE3MjUxNzc0NzksImV4cCI6MTcyNTE3NzU5OX0.aMx5SUlnyuDRo4tc3i9Ntx04nXLkPZQ5KDyWedVl88g
...
    GetAllUsers:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "adminpro@gmail.com"
        phone:
          type: string
          example: "0123456789"
        password:
          type: string
          format: password
          example: "$2b$10$WzI8V0rB0yqRLlUXM.JYuOBtS4KK2HyOvFzz7ma.eoF6w77qFO1eG"
        role_id:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2024-08-30T16:49:24.105Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-08-30T16:57:59.453Z"
...
    UsersResponse:
      type: array
      items:
        $ref: '#/components/schemas/GetAllUsers'
...
    TokenInvalidErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Token is not valid"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken